requirements.txt
absl-py==2.3.1
cvxopt==1.3.2
etils==1.13.0
fsspec==2025.10.0
glfw==2.10.0
imgui==2.0.0
importlib_resources==6.5.2
mujoco==3.3.7
numpy==2.2.6
opencv-python==4.12.0.88
PyOpenGL==3.1.10
pyzmq==27.1.0
scipy==1.15.3
typing_extensions==4.15.0
zipp==3.23.0

mujoco>=3.1.0

glfw>=2.6.0

imgui[glfw]>=2.0.0

numpy>=1.24.0
scipy>=1.11.0          # used for IK solver (scipy.optimize.minimize)

opencv-python>=4.8.0

pyzmq>=25.0.0